+++ R/imports.R
@@ -1,6 +1,8 @@
 #' @importFrom methods as S3Part<- setOldClass
+#' @importFrom rcrossref cr_cn
 #' @importFrom stats na.omit aggregate
 #' @importFrom stringr fixed regex str_count str_match_all str_replace str_trim
+#'     str_split
 #' @importFrom tools file_path_sans_ext
 #' @importFrom utils bibentry capture.output
 #' @importFrom yamlme render_rmd write_rmd