+++ examples/doi2bib.R
@@ -1,0 +1,6 @@
+## Retrieve bibliographic entries for two publications
+dois <- c("10.1111/1748-5967.12330",
+    "10.1016/j.envsci.2019.03.017")
+
+bibrefs <- doi2bib(dois)
+bibrefs